refactor(startup): split workspace startup controller
